Flaky reconnect tests - split up reconnect tests into multiple files - wait for 5 seconds to allow leader to be ready to accept the append - add option to test cluster to set log level
