+++ package.json
@@ -70,7 +70,7 @@
     ]
   },
   "dependencies": {
-    "@grpc/grpc-js": "^1.4.1",
+    "@grpc/grpc-js": "^1.4.5",
     "@types/debug": "^4.1.7",
     "@types/google-protobuf": "^3.15.5",
     "@types/node": "^16.11.4",
