+++ src/__test__/utils/index.ts
@@ -11,6 +11,7 @@ export * from "./collect";
 export * from "./Defer";
 export * from "./delay";
 export * from "./testEvents";
+export * from "./getCurrentConnection";
 export * from "./matchServerVersion";
 export * from "./optionalDescribe";
 export * from "./postEventViaHttpApi";
